+++ example/lib/pages/my_home_page_drawer.dart
@@ -1,3 +1,4 @@
+import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ncnn_yolox_flutter_example/providers/ncnn_yolox_options.dart';
@@ -26,6 +27,75 @@ class MyHomePageDrawer extends HookConsumerWidget {
                       );
             },
           ),
+          ListTile(
+            title: const Text('nmsThresh'),
+            subtitle: const Text('NMS threshold.'),
+            trailing: Text(ref.watch(ncnnYoloxOptions).nmsThresh.toString()),
+            onTap: () async {
+              final results = await showTextInputDialog(
+                context: context,
+                title: 'nmsThresh',
+                textFields: [
+                  const DialogTextField(
+                    keyboardType: TextInputType.numberWithOptions(
+                      decimal: true,
+                    ),
+                  ),
+                ],
+              );
+
+              ref.read(ncnnYoloxOptions.notifier).state =
+                  ref.read(ncnnYoloxOptions).copyWith(
+                        nmsThresh: double.parse(results!.first),
+                      );
+            },
+          ),
+          ListTile(
+            title: const Text('confThresh'),
+            subtitle: const Text('Threshold of bounding box prob.'),
+            trailing: Text(ref.watch(ncnnYoloxOptions).confThresh.toString()),
+            onTap: () async {
+              final results = await showTextInputDialog(
+                context: context,
+                title: 'confThresh',
+                textFields: [
+                  const DialogTextField(
+                    keyboardType: TextInputType.numberWithOptions(
+                      decimal: true,
+                    ),
+                  ),
+                ],
+              );
+
+              ref.read(ncnnYoloxOptions.notifier).state =
+                  ref.read(ncnnYoloxOptions).copyWith(
+                        confThresh: double.parse(results!.first),
+                      );
+            },
+          ),
+          ListTile(
+            title: const Text('targetSize'),
+            subtitle: const Text(
+              'Target image size after resize, might use 416 for small model.',
+            ),
+            trailing: Text(ref.watch(ncnnYoloxOptions).targetSize.toString()),
+            onTap: () async {
+              final results = await showTextInputDialog(
+                context: context,
+                title: 'targetSize',
+                textFields: [
+                  const DialogTextField(
+                    keyboardType: TextInputType.number,
+                  ),
+                ],
+              );
+
+              ref.read(ncnnYoloxOptions.notifier).state =
+                  ref.read(ncnnYoloxOptions).copyWith(
+                        targetSize: int.parse(results!.first),
+                      );
+            },
+          )
         ],
       ),
     );