+++ example/lib/providers/ncnn_yolox_options.dart
@@ -1,5 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:ncnn_yolox_flutter/ncnn_yolox.dart';
 
 part 'ncnn_yolox_options.freezed.dart';
 
@@ -7,6 +8,9 @@ part 'ncnn_yolox_options.freezed.dart';
 class NcnnYoloxOptions with _$NcnnYoloxOptions {
   const factory NcnnYoloxOptions({
     @Default(true) bool autoDispose,
+    @Default(yoloxNmsThreshDefault) double nmsThresh,
+    @Default(yoloxConfThreshDefault) double confThresh,
+    @Default(yoloxTargetSizeDefault) int targetSize,
   }) = _NcnnYoloxOptions;
 }